+++ InformationProtection/functions/Connect-InformationProtection.ps1
@@ -49,6 +49,11 @@
 		The tenant ID of the Entra application to use to authenticate.
 		Defaults to: "organizations" (Which means the tenant, the selected account belongs to)
 
+	.PARAMETER Certificate
+		The certificate to use for authenticating to Entra for the AzureRightsManagement and MIPSyncService services.
+		Requires specifying both ClientID and TenantID.
+		When used, this will first authenticate as application to the provided application, before registering those sessions for the MIP SDK's use.
+
 	.PARAMETER Email
 		Email address to register on the session object.
 		Used for metadata when labelling files.
@@ -84,6 +89,9 @@
 		[string]
 		$TenantID = 'organizations',
 
+		[System.Security.Cryptography.X509Certificates.X509Certificate2]
+		$Certificate,
+
 		[string]
 		$Email,
 
@@ -97,8 +105,16 @@
 		$services = $script:_serviceSelector.GetServiceMap($ServiceMap)
 
 		if ($ClientID) {
-			Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement
-			Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.MIPSyncService -UseRefreshToken
+			if (-not $Certificate) {
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.MIPSyncService -UseRefreshToken
+			}
+			else {
+				if (-not ($TenantID -as [guid])) {
+					Stop-PSFFunction -String 'Connect-InformationProtection.Error.NoTenantId' -Cmdlet $PSCmdlet -EnableException $true -Category InvalidArgument
+				}
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement, $services.MIPSyncService -Certificate $Certificate
+			}
 		}
 
 		Assert-EntraConnection -Cmdlet $PSCmdlet -Service $services.AzureRightsManagement